requirements.txt
# COCO Dataset Preparation for CLIP Fine-tuning
# Requirements for ELEC 475 Lab 4

# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0

# Transformers for CLIP
transformers>=4.30.0

# Dataset and Utilities
kaggle>=1.5.16
pycocotools>=2.0.6

# Image Processing
Pillow>=9.0.0

# Visualization and Progress
matplotlib>=3.5.0
tqdm>=4.65.0

# Additional utilities
numpy>=1.23.0
